v4/go: fix close-gate re-review residue and 32-bit cross-build (SOW-0025)
The Aspect-1 and Aspect-2 re-reviews of the previous gate batch returned
verifiable residue; each parity item was checked against the Rust source
before editing:

- the missing-membership ID corrupt detail now matches Rust
  stored_word_count wording, and the unreachable inline-record
  lost-location branch is removed (membership_dictionary.go).
- the membership blob leaf walk carries the expected branch level as a
  value plus set flag instead of allocating a per-level *uint16,
  matching Rust Option<u16> by value (membership_blob.go).
- the feed catalog lookup charges the tree-lookup counter before the
  absent-root shortcut like Rust fixed_tree::query, and the blob branch
  writer propagates the raw push error instead of wrapping it
  (feed_catalog.go, membership_blob.go).
- the history plan window-count check compares uint64 against
  MaxUint32, fixing the linux/386 and linux/arm cross-builds
  (history.go).
- the range-edit state machine's last pointer-Option residue (change,
  segment, writeReplacement, transform operation) now uses the
  optionalValue value+present shape already used by the ordered merge,
  removing the per-call &value escape (range_edit.go, test updated).
- the tree gap module's two byte-identical neighbor-cell structs are
  merged into one rejectCell type, matching Rust's single
  Option<(usize, R)> shape (gap.go).

Validation (all under nice): default and v4work suites (-count=1 and
-race), vet, checkptr=2, gofmt, six cross-builds including linux/386 and
linux/arm, and both allocation ceilings (39 and 84 per run) - all green.

+++ v4/go/internal/writer/history.go
@@ -152,7 +152,7 @@ func prepareHistoryPlan(store *DraftStore, windows []HistoryWindow, check func()
 	if windowCount == 0 {
 		return nil, invalid("history windows are empty")
 	}
-	if windowCount > math.MaxUint32 {
+	if uint64(windowCount) > math.MaxUint32 {
 		return nil, invalid("history window count exceeds u32")
 	}
 	heap := newHeapBudget(store.budget.MaxHeapBytes)
